change js function

+++ assets/js/index.js
@@ -128,8 +128,8 @@ $(document).ready(function () {
     costog = document.getElementById("cog").value;
 
     var saleprice = parseFloat(saleprice) || 0;
-    var costog = parseInt(costog) || 0;
-    var amazon_fee = parseInt(amazon_fee) || 0;
+    var costog = parseFloat(costog) || 0;
+    var amazon_fee = parseFloat(amazon_fee) || 0;
     var profit = saleprice - amazon_fee - costog;
     document.getElementById("profit").value = profit.toFixed(2)
   }
